requirements.txt
certifi==2019.6.16
Click==7.0
cycler==0.10.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.0
opencv-contrib-python==4.1.0.25
opencv-contrib-python-headless==4.1.0.25
Pillow==6.1.0
pkg-resources==0.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
six==1.12.0
torch==1.2.0+cpu
torchvision==0.4.0+cpu
Werkzeug==0.15.4
xmltodict==0.12.0
